+++ src/actions/file.ts
@@ -1,11 +1,12 @@
 import type { ITranslation } from "@/lib/types/data"
-import type { CreateTranslationResult, IBackendTranslation } from "@/lib/types/data/backend"
+import { XliffMetadata, type CreateTranslationResult, type IBackendTranslation } from "@/lib/types/data/backend"
 import { getErrorMessage } from "@/lib/utils"
-import { NewTranslationSchema, OpenTranslationSchema } from "@/schemas"
-import { NewTranslationType, OpenTranslationType } from "@/schemas/types"
+import { NewTranslationSchema, OpenTranslationSchema, OpenXliffSchema } from "@/schemas"
+import { NewTranslationType, OpenTranslationType, OpenXliffType } from "@/schemas/types"
 import { invoke } from "@tauri-apps/api/core"
 import { save } from "@tauri-apps/plugin-dialog"
-import { getFormatFromPath } from "@/lib/helpers"
+import { detectLanguageCode, getFormatFromPath } from "@/lib/helpers"
+import { ILangInputState } from "@/lib/types"
 
 export default class FileActions{
      public static async newTranslation(values: NewTranslationType): Promise<{
@@ -18,9 +19,11 @@ export default class FileActions{
                const validatedFields = NewTranslationSchema.safeParse(values)
                if(!validatedFields.success) return {error: "All fields are invalid", data: []}
                const {path, targetLanguageCode, format} = validatedFields.data
+               const meta = await invoke<XliffMetadata>("get_xliff_meta", { path })
                const res = await invoke<CreateTranslationResult>("create_translation", {
                     basePath: path,
-                    targetLanguageCode: targetLanguageCode,
+                    baseLang: !meta ? detectLanguageCode(path) : meta.src_lang,
+                    targetLang: targetLanguageCode,
                     format
                })
                return {
@@ -65,7 +68,31 @@ export default class FileActions{
                return {error: getErrorMessage(err), data: []}
           }
      }
-     public static async saveAll(table: ITranslation[], targetPath: string){
+     public static async openXliff(values: OpenXliffType) {
+          try {
+               const validatedFields = OpenXliffSchema.safeParse(values)
+               if(!validatedFields.success) return {error: "All fields are invalid", data: []}
+               const {translationPath} = validatedFields.data
+               const fileFormat = await getFormatFromPath(translationPath);
+               if(!fileFormat) return {error: "Format doesn't exist"}
+               const res = await invoke<IBackendTranslation[]>("open_xliff", {
+                    path: translationPath
+               })
+               return {
+                    success: "Translation opened successfully",
+                    data: res.map(val=>({
+                         keyName: val.key_name,
+                         translationString: val.translation_string,
+                         lineNumber: val.line_number,
+                         baseString: val.base_string
+                    }))
+               }
+          } catch (err) {
+               console.error(err)
+               return {error: getErrorMessage(err), data: []}
+          }
+     }
+     public static async saveAll(table: ITranslation[], targetPath: string, langs: ILangInputState){
           if(!targetPath) return {error: "Target Path is empty. Save cancelled"}
           if(table.length<=0) return {error: "Table itself is empty"}
           try {
@@ -78,7 +105,9 @@ export default class FileActions{
                await invoke("save_translation", {
                     targetPath,
                     entries,
-                    format: await getFormatFromPath(targetPath)
+                    format: await getFormatFromPath(targetPath),
+                    baseLang: langs.base,
+                    targetLang: langs.target
                })
                return {success: "Translation saved successfully"}
           } catch (err) {
@@ -86,7 +115,7 @@ export default class FileActions{
                return {error: getErrorMessage(err)}
           }
      }
-     public static async saveAs(table: ITranslation[]){
+     public static async saveAs(table: ITranslation[], langs: ILangInputState){
           if(table.length<=0) return {error: "Table itself is empty"}
           try {
                const targetPath = await save({
@@ -124,7 +153,9 @@ export default class FileActions{
                await invoke("save_translation", {
                     targetPath,
                     entries,
-                    format: await getFormatFromPath(targetPath)
+                    format: await getFormatFromPath(targetPath),
+                    baseLang: langs.base,
+                    targetLang: langs.target
                })
                return {success: "Translation saved successfully"}
           } catch (err) {